Actions: Use new folder design
This updates the action icons with the new
rounded-square folder design from upstream.

Additionally to upstream, 22px icons were added.
Removed folder action icons above 48px, these seem
not to be used (and likely nothing beyond 24px is used,
but still keeping 32px and 48x for now to be safe).